requirements.txt
neo4j>=5.24,<6
python-dotenv>=1.0,<2
openalex-official